Fix for Issue #38
Reverting to non fused implementation of msub will cause performance issues as msub is used heavily in intersection kernels. aarch64 does not have an equal to x86_64 msub, to fix the issue we:

reverted msub to original behaviour
changed msub usages in the intersection kernels to use madd with negation of the pre calculated constant vale (org_rdir).
all tests pass and performance is equal/better to what was before (depends on compiler)